Transaction result blocking send
This fixes a race condition where a transaction result could attempt to
send on the result channel before it had a receiver (i.e. before
transaction.WaitForResult has been called), causing the caller to never
see the result. This was very likely to occur when when two peers and
the STUN/TURN server were on the same host.