Close accepted TURN listener connections
Previously, connections accepted from the provided net.Listener would
never be closed, even if there was an error.

I had to refactor the NewServer method a bit to avoid hitting the
gocognit linter. The only real changes I made were to the connection
handler goroutine in the net.Listener goroutine.