UPDATE UTIL.GO. MAKE TYPE ASSERTION OPTIONAL
The test functions use the net library internal functions to create
examples of UDPAddr and TCPAddr for interpretation.

Provide unit tests for the new functions to validate the interpretation
of UDP and TCP addresses and ports

Provide an example of the mocking that now becomes possible. Whilst the
"typecast" appears neat it prevents test code from providing it's own
mocked contents from the Network() and String() functions that net.Addr
requires.

This means testing code cannot easily spoof IP addresses or
ports of interest. It severely limits the ability to test TURN code in
isolation as the server internals rely on this function - and link them
to the request packet and destination address.