Ignore "use of closed network connection" errors
Every time the server returns a response to the client it calls
BuildAndSend(), which assumes that the other side is still listening
and returns a "use of closed network connection" error when it does
not. This error always shows up when the client disconnects first,
which is totally normal, but messes up logs and fires stale
alarms. This commit silences these errors.

I am fairly certain that this change does not break anything. Yet,
please give feedback.