Fix reused loop counter in the server benchmark
RunBenchmarkServer uses two embedded loops to stress-test the server:
an outer loop to make just as many iterations as required for the
requested benchmark runtime, and an inner loop that walks through the
clients and sends one package from each. Formerly, both loops used the
same loop counter variable which led to incorrect results. This change
fixes this issue.