Add a quota handler callback
This PR adds a quota handler callback function which, if specified, is
called by the server just before making an allocation for a user. The
handler should return a single bool: if true then the allocation
request can proceed, otherwise the request is rejected with a 486
(Allocation Quota Reached) error.